base of chat, middleware, jwt refresh, improve better frontend module, small fixes
Note: not finished yet.
Implementation of websockets to manage the chat and the game (each one will have its own socket since it is easier to work this way and the code from one will not be mixed with the other), changes in the html structure to be able to do the more modular index page.
Implementation of a generic socket class in the frontend, and two separate ones for the game and chat instances. Chat implementation is ongoing, currently only connecting users appear and disappear (excluding yourself).
I have added the update of the jwt token, (I include it in this same branch since it was a minor change and is linked to the socket responses in addition to the http ones)