Almost done
Installation of ganache and truffle.
Ganache is an environment that provides a local blockchain, in our case for etherium.
Truffle is a development environment that allows you to compile and deploy contracts.
How it works: In docker, ganache is installed, which is a service that runs, this being our blockchain. On the other hand, a service (truffle) is started from docker, which is the compiler of the contract and the one in charge of deploying it in the blockchain, this is in charge of executing the migrations and then the container is turned off.
Disadvantages in development: when using docker, ganache temporarily uses a network which is eliminated once the container is turned off, to solve this I have created a volume to store it and that every time the container is raised it connects to the existing network.
On the other hand, truffle works "similarly" to django, where it has a migration file that detects changes to recompile the contract and deploy it. According to the documentation, the contract should not be deployed if there have been no changes, but this was not the case, in each truffle migrate the contract was deployed giving a new contract address, causing the data to be lost by staying in the old contract , I have solved it by checking whether or not the contract's .sol file has had changes to launch truffle migrate or not.
In the django part, I have created two GET and POST functions in views which wait to receive a json (the format is at the top in a comment) which is responsible for storing the blockchain data by calling the contract, and another function that returns the stored tournaments. For the store function, it is probably better to have it executed without the need for a message, and for only the server to call said function at the end of a tournament.

IMPORTANT NOTE: Everything is not finished yet, for security reasons the transfer of the owner of the contract needs to be implemented so that the functions can only be called by the server.