Cosas varias (#48)
* Almost done

Installation of ganache and truffle.
Ganache is an environment that provides a local blockchain, in our case for etherium.
Truffle is a development environment that allows you to compile and deploy contracts.
How it works: In docker, ganache is installed, which is a service that runs, this being our blockchain. On the other hand, a service (truffle) is started from docker, which is the compiler of the contract and the one in charge of deploying it in the blockchain, this is in charge of executing the migrations and then the container is turned off.
Disadvantages in development: when using docker, ganache temporarily uses a network which is eliminated once the container is turned off, to solve this I have created a volume to store it and that every time the container is raised it connects to the existing network.
On the other hand, truffle works "similarly" to django, where it has a migration file that detects changes to recompile the contract and deploy it. According to the documentation, the contract should not be deployed if there have been no changes, but this was not the case, in each truffle migrate the contract was deployed giving a new contract address, causing the data to be lost by staying in the old contract , I have solved it by checking whether or not the contract's .sol file has had changes to launch truffle migrate or not.
In the django part, I have created two GET and POST functions in views which wait to receive a json (the format is at the top in a comment) which is responsible for storing the blockchain data by calling the contract, and another function that returns the stored tournaments. For the store function, it is probably better to have it executed without the need for a message, and for only the server to call said function at the end of a tournament.

IMPORTANT NOTE: Everything is not finished yet, for security reasons the transfer of the owner of the contract needs to be implemented so that the functions can only be called by the server.

* Update views.py

* Update compose.yml

* Update compose.yml

* include migrations

* Only owner can add tournaments results

* add bridge mode to allow comunication between truffle and ganache

* fix

* Bug fixes (read description) and implementations

-When a game window such as billiards was opened, and then another pong window was opened and closed, billiards was disconnected since the socket connection was closed.
-The chat did not correctly display the number of unread messages in private chats
-When billiards was being played and the pong window was opened, the billiards canvas moved down through the css that included the canvas and not a specific id
-Added a view in the profile to be able to request the profile of another userid
-From the chat you can now open another user's profile at the request of the subject
-Corrected problem that when having two game instances open, the server received the game inputs from both causing the disconnection by mistake, now it is sent which game the input comes from and the server derives the input received to the corresponding game.

* Update views.py

* save

* last changes

* rm

* Basic cli

* fix

* 2fa

* navigation backward and forward fix

* Update router.js

* tournament table fix

* send table with winner

* Tournaments are finally saved on the blockchain 💯

* fix for points counter, if user disconnect before send "player ready" and not reconnect in time, rival win

* notify rounds tournament in chat

* fix in merge

* Include CLI

* fix

* fix to notification works with middleware changes

* improve better cli and documentation

* Delete cli directory

* fix in router, split pong and pool points

* Delete README.md

* rename

* Update README.md

* 3d pong, added option to play with algorithm, fix django crash if redis not start, rename compose to docker-compose

* Toggle view upon ending local games

* Points to win back to 6

* Solve bug in local multiplayer, second player up input wasnt handle correctlly

* save

* Improve profile, few stuff are fixed

* pong countdown, pong line, dix scrollbar inside game

* Update docker-compose.yml

* Update consumers.py

* up

* Revert "up"

This reverts commit 0e7b548367163658dc520afc8d0a2a0da43c3911.

* changes in friends UI, fix tournaments error after merge, fixed scroll after merge, now all window have unique css and not affect elements outside window

* algin buttons, fix auto css prefix

* loval variables used instead hardcoded, prevent frame skip on reset

* split table in a separate file to reutilice code

* shell

---------

Co-authored-by: RobertoRobrodri <karmmero@gmail.com>